% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_location_history.R
\name{clean_location_history}
\alias{clean_location_history}
\title{Clean Patient Location History Data}
\usage{
clean_location_history(
  data,
  uniqueID,
  startDate,
  endDate,
  unitName = NULL,
  roomNum = NULL
)
}
\arguments{
\item{data}{A \code{\link{data.frame}} containing patient location history data}

\item{uniqueID}{A unique identifier for each encounter or patient}

\item{startDate}{Column name for the column containing the start date for an enounter}

\item{endDate}{Column name for the column containing the end date for an enounter}

\item{unitName}{Optional column name for the column containing the unit name data}

\item{roomNum}{Optional column name for the column containing the room number data}
}
\value{
A \code{\link{data.frame}} containing the cleaned patient location history data
}
\description{
To use other location history functions in this package, your data must have columns for: patient's first unit and patient's next unit or patient's first room and patient's next room. There might be multiple rows per patient, but each row should always have a beginning and ending room and/or unit for each encounter.
}
