% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_expose_data.R
\docType{data}
\name{disease_expose_data}
\alias{disease_expose_data}
\title{Sample Disease-Exposure Data}
\format{
A \code{data.frame} with 5 columns and 500 rows:
\describe{
\item{id}{integer of the row id, could be thought of as the person id}
\item{disease}{character column specifying the disease status, with "No" equal to not diseased and "Yes" equal to diseased}
\item{exposure1}{binary column specifying if the person has/had the exposure (1) or not (0)}
\item{exposure2}{character column specifying if the person has/had the exposure ("Yes") or not ("No")}
\item{exposure3}{character column specifying which category of an exposure/risk factor the person falls into (are they a child, office worker, shop worker, etc.)}
}
}
\usage{
disease_expose_data
}
\description{
A simple example of a \code{data.frame} of disease-exposure/disease-risk factor data that could be used with this package and is used in examples
}
\keyword{datasets}
