% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.disease_expose.R
\name{plot.disease_expose}
\alias{plot.disease_expose}
\title{Plot Data for disease_expose Object}
\usage{
\method{plot}{disease_expose}(x, ...)
}
\arguments{
\item{x}{an object of class \code{disease_expose}, created with \code{\link{disease_expose}}}

\item{...}{additional arguments to be passed to \code{\link{geom_bar}}, such as \code{position}}
}
\value{
A \code{\link{gtable}} object of grouped bar charts on one page
}
\description{
Create grouped bar charts for each disease-exposure combination on one page
}
