% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_disease_expose.R
\name{clean_disease_expose}
\alias{clean_disease_expose}
\title{Disease-Exposure Cleaning Function}
\usage{
clean_disease_expose(data, disease, noDisease = NULL, exposures)
}
\arguments{
\item{data}{The \code{\link{data.frame}} containing disease-exposure data}

\item{disease}{Quoted name or unquoted numeric value of the column containing the disease status data}

\item{noDisease}{Optional, with a default of NULL. A quoted string specifying how disease status equal to not diseased is listed in the disease column.}

\item{exposures}{The column(s) containing exposure data, either given as a vector of quoted names or unquoted numeric values or a single quoted name or unquoted numeric value (if there is only one exposure)}
}
\value{
A data.frame containing binary variables for disease and exposure data
}
\description{
Takes disease-exposure data and makes all columns binary, which is needed to use other functions in the package
}
\examples{
de_data <- disease_exposure_data
cleaned_de_data <- clean_disease_expose(data = de_data, disease = "disease", noDisease = "No", exposures = c("exposure1", "exposure2", "exposure3"))

cleaned2 <- clean_disease_expose(data = de_data, disease = "disease", noDisease = "No", exposures = c(3,4,5))

cleaned3 <- clean_disease_expose(data = de_data, disease = 2, exposures = c("exposure1", "exposure2", "exposure3"))
}
