% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_expose.R
\name{disease_expose}
\alias{disease_expose}
\title{Helper to Create a \code{disease_expose} Object}
\usage{
disease_expose(x)
}
\arguments{
\item{x}{The \code{\link{data.frame}} containing disease-exposure data, with all disease and exposure columns as binary variables}
}
\value{
An object of class \code{disease_expose} and \code{data.frame} containing binary variables for disease and exposure data
}
\description{
Creates an object of class \code{disease_expose} with the data provided and selected in the Shiny gadget
}
\examples{
\dontrun{

de_data <- disease_expose_data # use the example data in the package

cleaned_de_data <- clean_disease_expose(data = de_data, disease = "disease", noDisease = "No", exposures = c("exposure1", "exposure2", "exposure3")) # clean the data using specific columns in the dataset

de_object <- disease_expose(cleaned_de_data) # simply call the function and select specifics in the Shiny gadget that pops up
}
}
