% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_expose.R
\name{disease_expose}
\alias{disease_expose}
\title{Create a 'disease_expose' Object}
\usage{
disease_expose(x)
}
\arguments{
\item{x}{The \code{\link{data.frame}} containing disease-exposure data}
}
\value{
An object of class disease_expose and data.frame containing binary variables for disease and exposure data
}
\description{
Creates an object of class disease_expose with the data provided and selected in the Shiny gadget
}
