% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_disease_expose.R
\name{new_disease_expose}
\alias{new_disease_expose}
\title{Constructor to Create a 'disease_expose' Object}
\usage{
new_disease_expose(x, disease, exposures)
}
\arguments{
\item{x}{The \code{\link{data.frame}} containing disease-exposure data}

\item{disease}{Quoted name or unquoted numeric value of the column containing the disease status data}

\item{exposures}{The column(s) containing exposure data, either given as a vector of quoted names or unquoted numeric values or a single quoted name or unquoted numeric value (if there is only one exposure)}
}
\value{
An object of class disease_expose and data.frame containing binary variables for disease and exposure data
}
\description{
Creates an object of class disease_expose with the data provided
}
\examples{
de_data <- disease_expose_data
cleaned_de_data <- clean_disease_expose(data = de_data, disease = "disease", noDisease = "No", exposures = c("exposure1", "exposure2", "exposure3"))
de_object <- new_disease_expose(cleaned_de_data, disease = 1, exposures = 2:8)
}
