% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_hosp_network.R
\name{new_hosp_network}
\alias{new_hosp_network}
\title{Create an Object of Class hosp_network}
\usage{
new_hosp_network(
  x,
  fromUnit = NULL,
  toUnit = NULL,
  fromRoom = NULL,
  toRoom = NULL
)
}
\arguments{
\item{x}{A \code{\link{data.frame}} containing the patient location history data}

\item{fromUnit}{Optional, unquoted column name from \code{x} containing the from unit data}

\item{toUnit}{Optional, unquoted column name from \code{x} containing the to unit data}

\item{fromRoom}{Optional, unquoted column name from \code{x} containing the from room data}

\item{toRoom}{Optional, unquoted column name from \code{x} containing the to room data}
}
\value{
An object of class hosp_network
}
\description{
Create an Object of Class hosp_network
}
