% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.disease_expose.R
\name{plot.disease_expose}
\alias{plot.disease_expose}
\title{Plot Data for disease_expose Object}
\usage{
\method{plot}{disease_expose}(x, ...)
}
\arguments{
\item{x}{an object of class \code{disease_expose}, created with \code{\link{disease_expose}}}

\item{...}{additional arguments to be passed to \code{\link{geom_bar}}, such as \code{position}}
}
\value{
A \code{\link{gtable}} object of grouped bar charts on one page
}
\description{
Create grouped bar charts for each disease-exposure combination on one page
}
\examples{
de_data <- disease_expose_data # use the example data in the package

cleaned_de_data <- clean_disease_expose(data = de_data, disease = "disease", noDisease = "No", exposures = c("exposure1", "exposure2", "exposure3")) # clean the data using specific columns in the dataset

\dontrun{
de_object <- disease_expose(cleaned_de_data) # the best way to create a disease_expose object is by using the helper and selecting your choices in the Shiny gadget
}

de_object <- new_disease_expose(cleaned_de_data, disease = 1, exposures = 2:8) # another way to create a disease_expose object is to use the constructor and manually enter the information

plot(x = de_object) # plot your disease_expose object

plot(x = de_object, position = 'dodge') # provide an input to geom_bar using the ... argument to customize the plot a little more

}
