% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hosp_network.R
\name{plot.hosp_network}
\alias{plot.hosp_network}
\title{Plot a hosp_network Object}
\usage{
\method{plot}{hosp_network}(
  x,
  by = c("unit", "room"),
  type = c("simple", "hub score", "authority score"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{hosp_network}}

\item{by}{Two options, either \code{'unit'} or \code{'room'}, telling whether to graph by unit or room interconnectivity}

\item{type}{Three options, either \code{'simple'}, \code{'hub score'}, or \code{'authority score'}, telling how the nodes should be sized. Simple graphs the nodes as all the same size, while the other two graph the nodes with sizes corresponding to their respective scores (see \code{\link{hub_score}} and \code{\link{authority_score}}.}

\item{...}{Additional arguments to be passed to \code{\link{plot.igraph}}, such as \code{vertex.color}. The ones already set are listed in the details above.}
}
\value{
Returns NULL, invisibly, as the \code{\link{plot.igraph}} is used to output the plot
}
\description{
Plot a hosp_network Object
}
\details{
\code{edge.arrow.size}, \code{edge.curved}, \code{vertex.label}, \code{vertex.label.cex}, \code{vertex.size}, \code{asp}, \code{layout}, and \code{main} are the \code{\link{plot.igraph}} options already given by this function, so to do more customization, please choose other options.
}
