% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.disease_expose.R
\name{summary.disease_expose}
\alias{summary.disease_expose}
\title{Summarize Disease-Exposure Data}
\usage{
\method{summary}{disease_expose}(object, ...)
}
\arguments{
\item{object}{an object of class \code{disease_expose}, created with \code{\link{disease_expose}}}

\item{...}{additional arguments to be passed to \code{\link{rbindlist}}, such as \code{fill}}
}
\value{
A \code{\link{data.frame}} object containing statistics for each disease-exposure combination
}
\description{
Summarize disease-exposure data by finding various statistics (incidence, risk ratio, odds ratio, Fisher p-value, Chi-square p-value) for each disease-exposure combination.
}
