% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.hosp_network.R
\name{summary.hosp_network}
\alias{summary.hosp_network}
\title{Create a Summary of a \code{hosp_network} Object}
\usage{
\method{summary}{hosp_network}(object, by = c("unit", "room"), ...)
}
\arguments{
\item{object}{An object of class \code{hosp_network}}

\item{by}{Two options, either \code{'unit'} or \code{'room'}, telling whether to summarize the network by unit or room interconnectivity}

\item{...}{Additional arguments to be passed to \code{\link{hub_score}} and \code{\link{authority_score}}, such as \code{scale}. The one already set is listed in the details section.}
}
\value{
A list containing the edges, nodes, largest cliques, hub score, authority score, edge density, mean distance, betweenness, reciprocity, degree, Eigen centrality, and assortativity degree of the network given in the object. Check out the following help pages to learn more about the functions used to calculate statistics:
\itemize{
  \item Largest Cliques - \code{\link{largest_cliques}}
  \item Hub Score - \code{\link{hub_score}}
  \item Authority Score - \code{\link{authority_score}}
  \item Edge Density - \code{\link{edge_density}}
  \item Mean Distance - \code{\link{mean_distance}}
  \item Betweenness - \code{\link{betweenness}}
  \item Reciprocity - \code{\link{reciprocity}}
  \item Degree - \code{\link{degree}}
  \item Eigen Centrality - \code{\link{eigen_centrality}}
  \item Assortativity Degree - \code{\link{assortativity_degree}}
}
}
\description{
Summarize a \code{hosp_network} object by finding various statistics (hub score, authority score, Eigen centrality, etc.) for the network.
}
\details{
\code{weights} is the \code{\link{hub_score}} and \code{\link{authority_score}} option already given in this function, so to do more customization, please choose other options.
}
\section{Thanks}{

Special thanks to Josh Sadowski for providing ideas of statistics to use for this function.
}

\examples{
hn_data <- hosp_network_data # read in example data provided in package

cleaned_hn_data <- clean_hosp_network(data = hn_data, uniqueID = UniqueEncountID, startDate = BeginDate, endDate = EndDate, unitName = UnitName, roomNum = RoomNumber) # clean the data using specific columns in the dataset

hn_object <- hosp_network(x = cleaned_hn_data, fromUnit = UnitName, toUnit = next_unit, fromRoom = RoomNumber, toRoom = next_room) # create an object of class hosp_network

summary(hn_object) # create summary for hosp_network object's unit data

summary(hn_object, by = "room", scale = FALSE) # create summary for hosp_network object's room data but do not scale hub scores and authority scores
}
